+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 All Notable changes to `oauth2-eve` will be documented in this file
 
+## 0.2.0 - 2016-02-04
+
+### Added
+- `Killmails\OAuth2\Client\Provider\EveOnline::getCharacterID`
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Complete code coverage
+
+### Removed
+- `Killmails\OAuth2\Client\Provider\EveOnline::getCharacterId`
+
+### Security
+- Nothing
+
 ## 0.1.0 - 2016-02-03
 
 ### Added